+++ plugins/alignments/src/LinearAlignmentsDisplay/components/useAlignmentsBase.ts
@@ -101,8 +101,8 @@ export interface LinearAlignmentsDisplayModel {
       number,
       {
         regionStart: number
-        arcX1: Float32Array
-        arcX2: Float32Array
+        arcX1: Uint32Array
+        arcX2: Uint32Array
         arcColorTypes: Float32Array
         arcIsArc: Uint8Array
         numArcs: number
